Move WordSegmenter into word segmentation package
